+++ lib/positioning.rb
@@ -1,4 +1,5 @@
 require_relative "positioning/version"
+require_relative "positioning/advisory_lock"
 require_relative "positioning/mechanisms"
 
 require "active_support/concern"
@@ -41,6 +42,11 @@ module Positioning
             send :"#{column}_will_change!"
             super(position)
           end
+
+          before_save { AdvisoryLock.new(self).aquire }
+          before_destroy { AdvisoryLock.new(self).aquire }
+          after_commit { AdvisoryLock.new(self).release }
+          after_rollback { AdvisoryLock.new(self).release }
 
           before_create { Mechanisms.new(self, column).create_position }
           before_update { Mechanisms.new(self, column).update_position }